update readme messages

+++ src/pipecraft-core/service_scripts/ORFfinder.sh
@@ -1,6 +1,14 @@
 #!/bin/bash
 
 #Using ORFfinder to identify open reading frames.
+# Input = fasta file.
+
+##################################
+###Third-party applications:
+#ORFfinder
+#seqkit
+#pigz
+##################################
 
 regex='[^/]*$'
 rep_seqs=${rep_seqs_file}
@@ -13,72 +21,82 @@ start_codon=${start_codon}
 ignore_nested=${ignore_nested}
 strand=${strand}
 
-printf "Settings: 
-rep_seqs_file=$rep_seqs_file
-min_len=$min_len
-genetic_code=$genetic_code
-start_codon=$start_codon
-ignore_nested=$ignore_nested
-strand=$strand
-"
+#output dir
+output_dir=$"/input/"
+# Source for functions
+source /scripts/submodules/framework.functions.sh
 
 #############################
 ### Start of the workflow ###
 #############################
 start=$(date +%s)
-
-#output dir
-output_dir=$"/input/"
+# Check if rep_seqs is fasta
+extension=$(echo $rep_seqs | (awk 'BEGIN{FS=OFS="."} {print $NF}';))
+printf "\n fileFormat = $fileFormat \n"
+if [[ $extension == "fasta" ]] || [[ $extension == "fa" ]] || [[ $extension == "fas" ]]; then
+    :
+else
+    printf '%s\n' "ERROR]: $extension formatting not supported here for 'rep seqs file'!
+Supported extensions: fasta, fas, fa.
+>Quitting" >&2
+    end_process
+fi
 
 ### Using ORFfinder to identify open reading frames
-printf "Using ORFfinder to identify open reading frames ... \n"
-
-ORFfinder -in $rep_seqs_file \
+checkerror=$(ORFfinder -in $rep_seqs_file \
     -ml $min_len \
     -g $genetic_code \
     -s $start_codon \
     -n $ignore_nested \
     -strand $strand \
     -outfmt 1 \
-    -out ORFfinder.temp
+    -out ORFfinder.temp 2>&1)
+check_app_error
 
 ### retain the longest ORF
   # sort seqs by length
-seqkit sort -l ORFfinder.temp \
+seqkit sort --quiet -l ORFfinder.temp \
     | sed -e 's/:.*//' | \
     sed -e 's/lcl|//' \
     > sort.temp
   # remove duplicate sequences (seqs with multiple ORFs). First seq will be kept
-seqkit rmdup -n sort.temp -w 0 > ORFs.fasta
+seqkit rmdup --quiet -n sort.temp -w 0 > ORFs.fasta
 
 #write putative pseudogenes to a file
 grep "^>" ORFs.fasta | sed -e 's/>//' > headers.temp
-seqkit grep -w 0 -v -f headers.temp $rep_seqs_file > NUMTs.fasta
-seqkit seq -n NUMTs.fasta > NUMTs_names.txt
+checkerror=$(seqkit grep --quiet -w 0 -v -f headers.temp $rep_seqs_file > NUMTs.fasta 2>&1)
+check_app_error
+checkerror=$(seqkit seq --quiet -n NUMTs.fasta > NUMTs_names.txt 2>&1)
+check_app_error
 
+# count outputs
 input_seqs=$(grep -c "^>" $rep_seqs_file)
 numts=$(grep -c "^>" NUMTs.fasta)
 orfs=$(grep -c "^>" ORFs.fasta)
-
-printf "
-input = $input_seqs seqs;
-discarded $numts potential NUMTs;
-output = $orfs seqs\n\n"
 
 #################################################
 ### COMPILE FINAL STATISTICS AND README FILES ###
 #################################################
 printf "\nCleaning up and compiling final stats files ...\n"
-rm *.temp
+if [[ $debugger != "true" ]]; then
+    rm *.temp
+fi
 end=$(date +%s)
 runtime=$((end-start))
 
 #Make README.txt file for demultiplexed reads
-printf "ORFfinder run for identifying open reading frames.
+printf "Used ORFfinder to identify open reading frames (ORFs), kept the longest ORF sequence (see 'Core command' below for the used settings).
 
 # ORFs.fasta = representative sequences from the input file with ORFs (open reading frames); contains $orfs reads.
 # NUMTs.fasta = putative pseudogenes/off-target ORFs; contains $numts reads.
 # NUMTs_names.txt = list of OTU/ASV identifiers of putative pseudogenes/off-target ORFs.
+
+Core command -> 
+ORFfinder -in $rep_seqs_file_path -ml $min_len -g $genetic_code -s $start_codon -n $ignore_nested -strand $strand
+-ml = min_len
+-g = genetic_code (see https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi)
+-s = start_codon
+-n = ignore_nested
 
 Total run time was $runtime sec.
 
@@ -86,21 +104,16 @@ Total run time was $runtime sec.
 ###Third-party applications for this process [PLEASE CITE]:
 #ORFfinder v0.4.3 for finding ORFs
     #https://www.ncbi.nlm.nih.gov/orffinder/
-#seqkit v2.3.0 for sorting results
+#seqkit v2.3.0 for sorting ORFfinder results
     #citation: Shen W, Le S, Li Y, Hu F (2016) SeqKit: A Cross-Platform and Ultrafast Toolkit for FASTA/Q File Manipulation. PLOS ONE 11(10): e0163962. https://doi.org/10.1371/journal.pone.0163962
     #https://bioinf.shenwei.me/seqkit/
 ##################################################################" > $output_dir/README_ORFfinder.txt
 
 #Done
 printf "\nDONE"
-printf "Data in directory $output_dir\n"
-printf "Check README.txt file in $output_dir for further information about the process.\n"
 printf "Total time: $runtime sec.\n"
 
 #variables for all services
-echo "workingDir=/$output_dir"
+echo "workingDir=$output_dir"
 echo "fileFormat=$extension"
-echo "dataFormat=demultiplexed"
 echo "readType=single_end"
-
-
